refactor: move server-src into /backend
