+++ apps/nextjs-app/tsconfig.json
@@ -15,6 +15,7 @@
       "@/config/*": ["./config/*"],
       "@/lib/*": ["./lib/*"],
       "@/components/*": ["./components/*"],
+      "@/backend/*": ["./backend/*"],
       "@/_backend/*": ["./_backend/*"],
       "@/features/*": ["./features/*"],
       "@/pages/*": ["./pages/*"],
